refactor(P1.7): split assistant/tools/channel.py into package
The file was 646 LOC with 19 PydanticAI tool declarations in a single
register_channel_tools. Agents debugging assistant behavior had to scan the
whole file to locate a specific tool.

Split by concern:
- channels.py — list/add/edit/remove channel
- sources.py — get/add/remove RSS source
- pipeline.py — status/run/posts/cost/publish_text/generate_and_review
- schedule.py — posting schedule, per-post schedule/reschedule/cancel, publish schedule
- __init__.py — composes the four register_*_tools(agent) calls and
  re-exports _SCHEDULE_TIME_RE for test_assistant compatibility.

Kept: RunContext + AssistantDeps at runtime (noqa: TC001/TC002) — PydanticAI's
@agent.tool decorator calls typing.get_type_hints at registration, so TYPE_CHECKING
guards would NameError.

Tests: 640 passed, 2 skipped. Ruff + ty green.
